[refactor] change naming of ssd related file
