+++ SSD/file_iostream.h
