+++ SSD/ssd_nand.txt
@@ -1,0 +1,100 @@
+0 0x00000000
+1 0xEEEEAAAA
+2 0x00000000
+3 0x00000000
+4 0x00000000
+5 0x00000000
+6 0x00000000
+7 0x00000000
+8 0x00000000
+9 0x00000000
+10 0x00000000
+11 0x00000000
+12 0x00000000
+13 0x00000000
+14 0x00000000
+15 0x00000000
+16 0x00000000
+17 0x00000000
+18 0x00000000
+19 0x00000000
+20 0x00000000
+21 0x00000000
+22 0x00000000
+23 0x00000000
+24 0x00000000
+25 0x00000000
+26 0x00000000
+27 0x00000000
+28 0x00000000
+29 0x00000000
+30 0x00000000
+31 0x00000000
+32 0x00000000
+33 0x00000000
+34 0x00000000
+35 0x00000000
+36 0x00000000
+37 0x00000000
+38 0x00000000
+39 0x00000000
+40 0x00000000
+41 0x00000000
+42 0x00000000
+43 0x00000000
+44 0x00000000
+45 0x00000000
+46 0x00000000
+47 0x00000000
+48 0x00000000
+49 0x00000000
+50 0x00000000
+51 0x00000000
+52 0x00000000
+53 0x00000000
+54 0x00000000
+55 0x00000000
+56 0x00000000
+57 0x00000000
+58 0x00000000
+59 0x00000000
+60 0x00000000
+61 0x00000000
+62 0x00000000
+63 0x00000000
+64 0x00000000
+65 0x00000000
+66 0x00000000
+67 0x00000000
+68 0x00000000
+69 0x00000000
+70 0x00000000
+71 0x00000000
+72 0x00000000
+73 0x00000000
+74 0x00000000
+75 0x00000000
+76 0x00000000
+77 0x00000000
+78 0x00000000
+79 0x00000000
+80 0x00000000
+81 0x00000000
+82 0x00000000
+83 0x00000000
+84 0x00000000
+85 0x00000000
+86 0x00000000
+87 0x00000000
+88 0x00000000
+89 0x00000000
+90 0x00000000
+91 0x00000000
+92 0x00000000
+93 0x00000000
+94 0x00000000
+95 0x00000000
+96 0x00000000
+97 0x00000000
+98 0x00000000
+99 0x00000000
