+++ SSD/ssd_output.txt
@@ -1,0 +1,1 @@
+0xEEEEAAAA
